Add specific median method from numpy (lower by default)
This avoids a bug that would color a bin with a different color when the median is calculated over an even number of elements and the middle elements are different. For example, the median of [1, 1, 3, 3] would be 2. This has the effect of painting that bin with color number 2, which was never assigned by the user. To avoid this, we chose here to always get the lower median (1 in this example).