add `highlight_points_by_ids` public method
This allows one to highlight specific points in the scatter given some ids.
It doubles their sizes and make other points half transparent.
Passing an empty list or None to it resets sizes and alphas.
This could be used to link selected labels in a Labels layer or selected points in a Points layer to points in the Scatter artist.